SystemVerificationService: fall back to WMI to detect installed RAM
In case GetPhysicallyInstalledSystemMemory() fails, which seems to be the
case on some configurations with dynamically allocated video memory,
use a WMI query to determine the physically installed memory.

That is a better solution than calling e.g. GlobalMemoryStatusEx() because
that will exclude reserved memory that's not available to the OS.
But our intention is to check for installed memory, not available.