SystemVerificationService: ignore if device is not found
For USB devices which do not have unique IDs, `GetCurrentUSBDevice` is
unable to detect the USB device corresponding to the current disk drive.
This change causes the launcher to skip part of the compatibility check
if it is unable to detect the USB device corresponding to the connected
disk drive.

https://phabricator.endlessm.com/T30740