Allow to use launcher when PC is unable to restart to Endless
Allow to use the launcher (Open Kiwix and Kolibri, for example) even if the
computer is not able to boot Endless. Now a warning is show instead of the
button, allowing the user to get more information about it.

Another simplifications in the views were made, like removing the
ActionButton style from WelcomePage and reuse the DefaultButtonStyle.

WrongUSBPort page is removed, Incompatibility page is used for the
same purpose.

https://phabricator.endlessm.com/T30753