Remove VC++ runtime installation verification
If we install Kiwix 2.0.4 VC++ runtimes using the launcher and then get newer
Endless Key with updated Kiwix that includes newer VC++ runtimes, the launcher
won't install them since the launcher detects that the major version for that runtime
is already installed. But checking the major runtime is not enough.

For simplicity, let's not check at all and let decide the VC++ runtimes installer to
decide what to do. If the installer version it's already installed, the installer won't
do anything and silently exits, which is fine for us.

https://phabricator.endlessm.com/T31243